+++ src/DefaultConsumer.php
@@ -15,7 +15,7 @@ namespace PMG\Queue;
 use Psr\Log\LoggerInterface;
 
 /**
- * A default implementation of `Consumer`. Runs jobs via an executor.
+ * A default implementation of `Consumer`. Runs jobs via a MessageHandler.
  *
  * @since   2.0
  * @api
@@ -29,9 +29,9 @@ final class DefaultConsumer extends AbstractConsumer
     private $driver;
 
     /**
-     * @var MessageExecutor
+     * @var MessageHandler
      */
-    private $executor;
+    private $handler;
 
     /**
      * @var RetrySpec
@@ -40,13 +40,13 @@ final class DefaultConsumer extends AbstractConsumer
 
     public function __construct(
         Driver $driver,
-        MessageExecutor $executor,
+        MessageHandler $handler,
         RetrySpec $retries=null,
         LoggerInterface $logger=null
     ) {
         parent::__construct($logger);
         $this->driver = $driver;
-        $this->executor = $executor;
+        $this->handler = $handler;
         $this->retries = $retries ?: new Retry\LimitedSpec();
     }
 
@@ -63,9 +63,9 @@ final class DefaultConsumer extends AbstractConsumer
         $result = false;
         $message = $envelope->unwrap();
 
-        $this->getLogger()->debug('Executing message {msg}', ['msg' => $message->getName()]);
-        $result = $this->executeMessage($message);
-        $this->getLogger()->debug('Executed message {msg}', ['msg' => $message->getName()]);
+        $this->getLogger()->debug('Handling message {msg}', ['msg' => $message->getName()]);
+        $result = $this->handleMessage($message);
+        $this->getLogger()->debug('Handled message {msg}', ['msg' => $message->getName()]);
 
         if ($result) {
             $this->driver->ack($queueName, $envelope);
@@ -87,10 +87,10 @@ final class DefaultConsumer extends AbstractConsumer
         }
     }
 
-    private function executeMessage(Message $message)
+    private function handleMessage(Message $message)
     {
         try {
-            return $this->executor->execute($message);
+            return $this->handler->handle($message);
         } catch (Exception\MustStop $e) {
             // MustStop exceptions are thrown by handlers to indicate a
             // graceful stop is required. So we don't wrap them. Just rethrow
@@ -98,8 +98,8 @@ final class DefaultConsumer extends AbstractConsumer
         } catch (\Exception $e) {
             // any other exception is simply logged. We and marked as failed
             // below. We only log here because we can't make guarantees about
-            // the implementation of the executor and whether or not it actually
-            // throws exceptions on failure (see ForkingExecutor).
+            // the implementation of the handler and whether or not it actually
+            // throws exceptions on failure (see PcntlForkingHandler).
             $this->getLogger()->critical('Unexpected {cls} exception handling {name} message: {msg}', [
                 'cls'   => get_class($e),
                 'name'  => $message->getName(),