+++ CHANGELOG.md
@@ -18,6 +18,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   `NativeSerializer`'s constructor.
 - [BC BREAK] `AbstractPersistanceDriver::getSerializer` was removed, use
   `AbstractPersistanceDriver::assureSerializer` instead.
+- [BC BREAK] `Consumer::stop` now takes an optional exit code. Only really
+  relevant for implementors or the `Consumer` interface.
 - `Consumer` has docblocks that reflect its actual return values now.
 - `PheanstalkDriver` is no longer part of the core. Instead of requiring
    `pmg/queue` directly in your `composer.json`, require `pmg/queue-pheanstalk`
@@ -25,7 +27,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
-n/a
+- `DefaultConsumer` now catches and handles PHP 7's `Error` exceptions
 
 ### Added