+++ CHANGELOG.md
@@ -16,6 +16,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - [BC BREAK] `PMG\Queue\Serializer\SigningSerializer` has been merged into
   `NativeSerializer` and removed. Pass your key as the first argument to
   `NativeSerializer`'s constructor.
+- [BC BREAK] `AbstractPersistanceDriver::getSerializer` was removed, use
+  `AbstractPersistanceDriver::assureSerializer` instead.
+- [BC BREAK] `Consumer::stop` now takes an optional exit code. Only really
+  relevant for implementors or the `Consumer` interface.
+- [BC BREAK] `MessageExecutor`, `HandlerResolver`, and their implementations
+  have been removed. See `UPGRADE-3.0.md` for some info on migration.
 - `Consumer` has docblocks that reflect its actual return values now.
 - `PheanstalkDriver` is no longer part of the core. Instead of requiring
    `pmg/queue` directly in your `composer.json`, require `pmg/queue-pheanstalk`
@@ -23,7 +29,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
-n/a
+- `DefaultConsumer` now catches and handles PHP 7's `Error` exceptions
 
 ### Added