+++ CHANGELOG.md
@@ -3,7 +3,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## 3.0.0 (Unreleased)
+## 3.NEXT (Unreleased)
+
+### Changed
+n/a
+
+### Fixed
+n/a
+
+### Added
+n/a
+
+## 3.0.0
 
 ### Changed